# docs/swagger.yaml
basePath: /
definitions:
  entity.CheckResult:
    properties:
      check_name:
        type: string
      details:
        type: string
      raw_data: {}
      risk_found:
        type: boolean
      risk_level:
        $ref: '#/definitions/entity.RiskLevel'
      score_penalty:
        type: number
    type: object
  entity.PaymentStatus:
    enum:
    - pending
    - paid
    - canceled
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusPaid
    - PaymentStatusCanceled
  entity.RiskLevel:
    enum:
    - LOW
    - MEDIUM
    - HIGH
    - CRITICAL
    type: string
    x-enum-varnames:
    - RiskLevelLow
    - RiskLevelMedium
    - RiskLevelHigh
    - RiskLevelCritical
  payment.CreateInvoiceRequest:
    properties:
      address:
        example: 0x0000db5c8B030ae20308ac975898E09741e70000
        type: string
    required:
    - address
    type: object
  payment.CreateInvoiceResponse:
    properties:
      orderId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      orderUrl:
        example: https://t.me/pay?hash=1234567890abcdef
        type: string
    type: object
  payment.PaymentStatusResponse:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/entity.PaymentStatus'
        example: pending
    type: object
  payment.WebhookRequest:
    properties:
      message:
        properties:
          chat:
            properties:
              id:
                type: integer
              type:
                type: string
            type: object
          date:
            type: integer
          from:
            properties:
              id:
                type: integer
            type: object
          message_id:
            type: integer
          pre_checkout_query:
            properties:
              id:
                type: string
            type: object
          successful_payment:
            properties:
              currency:
                type: string
              invoice_payload:
                type: string
              provider_payment_charge_id:
                type: string
              telegram_payment_charge_id:
                type: string
              total_amount:
                type: integer
            type: object
        type: object
      update_id:
        type: integer
    type: object
  payment.WebhookResponse:
    properties:
      ok:
        type: boolean
    type: object
  wallet.CheckWalletRequest:
    properties:
      address:
        example: 0x0000db5c8B030ae20308ac975898E09741e70000
        type: string
      guid:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - address
    type: object
  wallet.CheckWalletResponse:
    properties:
      address:
        type: string
      checks:
        items:
          $ref: '#/definitions/entity.CheckResult'
        type: array
      errors:
        items:
          type: string
        type: array
      recommendations:
        type: string
      score:
        type: number
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for checking wallet security and calculating nutrition score
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Wallet Nutrition Score API
  version: "1.0"
paths:
  /api/check:
    post:
      consumes:
      - application/json
      description: Check wallet security and get nutrition score
      parameters:
      - description: Wallet address to check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wallet.CheckWalletRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.CheckWalletResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check wallet security
      tags:
      - wallet
  /api/payment-status/{guid}:
    get:
      consumes:
      - application/json
      description: Get payment status by GUID
      parameters:
      - description: Payment GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.PaymentStatusResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payment status
      tags:
      - payment
  /api/payment/create-invoice:
    post:
      consumes:
      - application/json
      description: Create a payment invoice for wallet check activation
      parameters:
      - description: Payment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.CreateInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.CreateInvoiceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create payment invoice
      tags:
      - payment
  /api/payment/webhook:
    post:
      consumes:
      - application/json
      description: Handle payment notifications from Telegram
      parameters:
      - description: Webhook data from Telegram
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.WebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.WebhookResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Handle Telegram webhook
      tags:
      - payment
  /health:
    get:
      consumes:
      - application/json
      description: Check if the service is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
swagger: "2.0"